requirements-ui.txt
streamlit>=1.36.0
ultralytics>=8.3.0
opencv-python>=4.10.0
pillow>=10.0.0
numpy>=1.26.0
